perf(catalog): split test_cli.py into 4 files for parallel CI distribution
With --dist=loadfile all 113 test functions ran on a single worker (438 s
wall). Splitting into test_cli_init_crud, test_cli_inspect_git,
test_cli_run_compose, and test_cli_params lets pytest-xdist spread the
load across four workers, targeting ~110 s tail instead of 438 s.

No fixture or test logic changed; all tests use fresh function-scoped
fixtures with no shared mutable state.
