+++ python/xorq/catalog/tests/test_cli_run_compose.py
@@ -1,0 +1,545 @@
+# Tests for: run, compose, compose+run roundtrip, --pdb flag, run-cached commands
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import pyarrow as pa
+import pytest
+from click.testing import CliRunner
+
+import xorq.api as xo
+from xorq.catalog.catalog import Catalog
+from xorq.catalog.cli import cli
+from xorq.cli import cli as top_cli
+from xorq.vendor.ibis.expr import operations as ops
+
+
+@pytest.fixture
+def runner():
+    yield CliRunner()
+
+
+@pytest.fixture
+def catalog_with_source_and_transform(catalog_path):
+    """Populate a catalog with a source entry and an unbound transform entry."""
+    catalog = Catalog.from_kwargs(path=catalog_path, init=False)
+
+    source = xo.memtable(
+        {"user_id": [1, 2, 3], "amount": [10.0, 20.0, 30.0], "name": ["a", "b", "c"]}
+    )
+    source_entry = catalog.add(source, aliases=("src",))
+
+    schema = source.schema()
+    unbound = ops.UnboundTable(name="placeholder", schema=schema).to_expr()
+    transform = unbound.filter(unbound.amount > 0).select("user_id", "amount")
+    transform_entry = catalog.add(transform, aliases=("trn",))
+
+    return catalog_path, source_entry.name, transform_entry.name
+
+
+# --- run command ---
+
+
+def test_run_single_entry(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "src", "-o", "-", "-f", "csv"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_run_default_output(runner, catalog_with_source_and_transform):
+    """Default output (no -o) writes to /dev/null, exits 0."""
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "src"],
+    )
+    assert result.exit_code == 0, result.output
+
+
+def test_run_with_limit(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "src", "-o", "-", "-f", "csv", "--limit", "1"],
+    )
+    assert result.exit_code == 0, result.output
+    lines = result.output.strip().splitlines()
+    assert len(lines) == 2  # header + 1 data row
+
+
+def test_run_with_code(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run",
+            "src",
+            "-c",
+            "source.filter(source.amount > 15)",
+            "-o",
+            "-",
+            "-f",
+            "csv",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+    # amount <= 15 should be filtered out (only 20.0 and 30.0 remain)
+    lines = result.output.strip().splitlines()
+    assert len(lines) == 3  # header + 2 data rows
+
+
+def test_run_rejects_unbound_entry(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "trn"],
+    )
+    assert result.exit_code != 0
+    assert "unbound expression" in result.output
+
+
+def test_run_two_entries(runner, catalog_with_source_and_transform):
+    """run src trn  — compose + execute in one shot."""
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "src", "trn", "-o", "-", "-f", "csv"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_run_two_entries_with_code(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run",
+            "src",
+            "-c",
+            "source.filter(source.amount > 15)",
+            "-o",
+            "-",
+            "-f",
+            "csv",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    lines = result.output.strip().splitlines()
+    assert len(lines) == 3  # header + 2 data rows
+
+
+def test_run_two_entries_json(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "src", "trn", "-o", "-", "-f", "json"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_run_two_entries_with_limit(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run",
+            "src",
+            "trn",
+            "-o",
+            "-",
+            "-f",
+            "csv",
+            "--limit",
+            "1",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    lines = result.output.strip().splitlines()
+    assert len(lines) == 2  # header + 1 data row
+
+
+def test_run_no_entries(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run"],
+    )
+    assert result.exit_code != 0
+
+
+def test_run_piped_arrow_into_unbound(runner, catalog_with_source_and_transform):
+    """run bound -o - -f arrow | run unbound  — simulated via CliRunner input."""
+
+    catalog_path, _, _ = catalog_with_source_and_transform
+
+    # Step 1: run the source entry, capture arrow bytes
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "src", "-o", "-", "-f", "arrow"],
+    )
+    assert result.exit_code == 0, result.output
+    arrow_bytes = result.output_bytes
+
+    # Step 2: pipe arrow bytes into the unbound entry
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "trn", "-o", "-", "-f", "csv"],
+        input=arrow_bytes,
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+    assert "amount" in result.output
+
+
+def test_run_nonexistent_entry(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "no-such-entry"],
+    )
+    assert result.exit_code != 0
+    assert "not found" in result.output
+
+
+# --- compose command ---
+
+
+def test_compose_two_entries(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "Cataloged as" in result.output
+
+
+def test_compose_with_alias(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "composed-result"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "Cataloged as" in result.output
+    assert "composed-result" in result.output
+
+
+def test_compose_with_code(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "compose",
+            "src",
+            "-c",
+            "source.filter(source.amount > 15)",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert "Cataloged as" in result.output
+
+
+def test_compose_code_no_entry(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "-c", "source.limit(1)"],
+    )
+    assert result.exit_code != 0
+
+
+def test_compose_no_entries(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose"],
+    )
+    assert result.exit_code != 0
+
+
+def test_compose_dry_run(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "--dry-run"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "Dry run" in result.output
+    assert "Schema:" in result.output
+    assert "user_id" in result.output
+    # dry-run should NOT catalog
+    assert "Cataloged" not in result.output
+
+
+def test_compose_without_alias_catalogs_by_hash(
+    runner, catalog_with_source_and_transform
+):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "Cataloged as" in result.output
+
+
+# --- compose + run roundtrip tests ---
+
+
+def test_compose_then_run_roundtrip(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    # compose with alias
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "my-composed"],
+    )
+    assert result.exit_code == 0, result.output
+    # run the composed entry
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "my-composed", "-o", "-", "-f", "csv"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_compose_then_run_json(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "json-test"],
+    )
+    assert result.exit_code == 0, result.output
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "json-test", "-f", "json", "-o", "-"],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_compose_then_run_parquet(runner, catalog_with_source_and_transform, tmpdir):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    out = str(Path(tmpdir).joinpath("out.parquet"))
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "pq-test"],
+    )
+    assert result.exit_code == 0, result.output
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "pq-test", "-f", "parquet", "-o", out],
+    )
+    assert result.exit_code == 0, result.output
+    assert Path(out).exists()
+
+
+def test_compose_then_run_csv_file(runner, catalog_with_source_and_transform, tmpdir):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    out = str(Path(tmpdir).joinpath("out.csv"))
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "csv-test"],
+    )
+    assert result.exit_code == 0, result.output
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "csv-test", "-f", "csv", "-o", out],
+    )
+    assert result.exit_code == 0, result.output
+    assert Path(out).exists()
+    assert "user_id" in Path(out).read_text()
+
+
+def test_compose_then_run_arrow_file(runner, catalog_with_source_and_transform, tmpdir):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    out = str(Path(tmpdir).joinpath("out.arrow"))
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "arrow-test"],
+    )
+    assert result.exit_code == 0, result.output
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "arrow-test", "-f", "arrow", "-o", out],
+    )
+    assert result.exit_code == 0, result.output
+    assert Path(out).exists()
+    with open(out, "rb") as f:
+        table = pa.ipc.open_stream(f).read_all()
+    assert len(table) > 0
+
+
+def test_compose_then_run_arrow_stdout(runner, catalog_with_source_and_transform):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "compose", "src", "trn", "-a", "arrow-stdout-test"],
+    )
+    assert result.exit_code == 0, result.output
+    result = runner.invoke(
+        cli,
+        ["--path", catalog_path, "run", "arrow-stdout-test", "-f", "arrow", "-o", "-"],
+    )
+    assert result.exit_code == 0, result.output
+    assert len(result.output) > 0
+
+
+# --- --pdb flag ---
+
+
+def test_pdb_flag_invokes_post_mortem(tmp_path, monkeypatch):
+    """--pdb should let exceptions propagate to PdbGroup, which calls pdb.post_mortem."""
+    mock_pm = MagicMock()
+    monkeypatch.setattr("xorq.cli.pdb.post_mortem", mock_pm)
+
+    # "catalog list" on a non-catalog directory fails inside the command body
+    # (not at Click arg-parsing time), so it exercises click_context_catalog.
+    runner = CliRunner()
+    result = runner.invoke(
+        top_cli,
+        ["--pdb", "catalog", "--path", str(tmp_path), "list"],
+    )
+    assert result.exit_code != 0
+    assert mock_pm.called, "pdb.post_mortem was not called with --pdb"
+
+
+def test_no_pdb_flag_wraps_exception(tmp_path):
+    """Without --pdb, errors should be wrapped as clean 'Error: ...' messages."""
+    runner = CliRunner()
+    result = runner.invoke(
+        top_cli,
+        ["catalog", "--path", str(tmp_path), "list"],
+    )
+    assert result.exit_code != 0
+    assert "Error:" in result.output
+
+
+# --- run-cached command ---
+
+
+def test_run_cached_single_entry(runner, catalog_with_source_and_transform, tmp_path):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    cache_dir = tmp_path / "cache"
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run-cached",
+            "src",
+            "--cache-dir",
+            str(cache_dir),
+            "-o",
+            "-",
+            "-f",
+            "csv",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+    assert cache_dir.exists()
+    assert any(cache_dir.iterdir())
+
+
+def test_run_cached_two_entries(runner, catalog_with_source_and_transform, tmp_path):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    cache_dir = tmp_path / "cache"
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run-cached",
+            "src",
+            "trn",
+            "--cache-dir",
+            str(cache_dir),
+            "-o",
+            "-",
+            "-f",
+            "csv",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_run_cached_snapshot_type(runner, catalog_with_source_and_transform, tmp_path):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    cache_dir = tmp_path / "cache"
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run-cached",
+            "src",
+            "--cache-type",
+            "snapshot",
+            "--cache-dir",
+            str(cache_dir),
+            "-o",
+            "-",
+            "-f",
+            "csv",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_run_cached_ttl(runner, catalog_with_source_and_transform, tmp_path):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    cache_dir = tmp_path / "cache"
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run-cached",
+            "src",
+            "--ttl",
+            "60",
+            "--cache-dir",
+            str(cache_dir),
+            "-o",
+            "-",
+            "-f",
+            "csv",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert "user_id" in result.output
+
+
+def test_run_cached_no_entries(runner, catalog_with_source_and_transform, tmp_path):
+    catalog_path, _, _ = catalog_with_source_and_transform
+    cache_dir = tmp_path / "cache"
+    result = runner.invoke(
+        cli,
+        [
+            "--path",
+            catalog_path,
+            "run-cached",
+            "--cache-dir",
+            str(cache_dir),
+        ],
+    )
+    assert result.exit_code != 0
+    assert "At least one entry is required" in result.output
